[BUGFIX] Fix Persistence\Doctrine\Query serialization cycle
If you assign a QueryResult to a ViewHelper and later on use
$result->getQuery(), this query is not executable any more, if
you used complex "where"-clauses involving joins, as these
joins are not reapplied to the object during the __wakeup()
call.

This patch ensures the query being callable like the
original one after __wakeup() by saving all joins in array
and reapplies them after wakeup. A functional test is
included.

Change-Id: I1e5a4b607738b2b4738e31ec68d1c2164201e115
Fixes: #31696
Releases: 1.0, 1.1