[BUGFIX] ACL file permissions do not work correctly
When setting file permissions the setting of ACL would succeed but
the set permissions did not work as expected. This changes fixes that.

In addition it makes the calls in a chain so that the check for the
return value of the chmod commands is more robust. The directories
that are created are now always owned by the command line user.

Change-Id: I19bb0acf0bdcf50b10ba3c3397c0a4d62e305e21
Fixes: #32499
Releases: 1.0, 1.1