[BUGFIX] Integrity constraint violation deleting resources
If you have two resources using the same resource pointer,
deleting a Resource will trigger deletion of the
connected ResourcePointer which then leads to an integrity
constraint violation.

This patch tells doctrine NOT to cascade delete operations.
This will lead to orphan resource-pointers which have to
be deleted manually or on a regular base.

Change-Id: I3c98bfc603d5093af96b991e26b9eff3b2f5c840
Fixes: #34448
Releases: 1.1, 1.2