[BUGFIX] Routing cache only works for one request
Currently the routing uses a TransientMemory cache backend by
default. Besides, objects in the routing values are converted
to hashes using spl_object_hash() - which generates a new hash
for every request.

This change configures the RouterCachingAspect to use a
SimpleFile cache backend and adjusts the aspect to use the
persistence manager in order to convert objects to their
identifiers before creating cache entries.

This also adds a signal to the FunctionalTestCase base
class that flushes the routing caches on teardown.

Furthermore this adds two convenience methods
getLastMatchedRoute() and getLastResolvedRoute() to the router
which should make debugging and testing easier.

Change-Id: I9605731f78cb1be9b0fb5c439d7149a676b6ccb7
Fixes: #38172
Releases: 1.1, 1.2