[!!!][BUGFIX] Fix request injection in security context
As inject* methods are excluded from the session
lazy loading proxy mechanism, the injectRequest
method of the security context didn't work correctly
in requests with a session timeout. Renaming the
method to setRequest solves the problem.

In case you implemented your own RequestHandler, this is
a breaking change, and you need to call setRequest() now.

Change-Id: Iaf12a23097433697e86897f4517caea1ef2b0b7d
Related: #37001
Releases: 1.1, 1.2