[!!!][BUGFIX] Expose timezone in JsonView's DateTime output
This corrects the string formatting in JsonView regarding
DateTime objects. The additionally required Timezone offset
is appended as of ISO 8601 in order to not lead into timezone
issues after transferring the resulting JSON.

This is breaking in cases where the original incorrect behaviour
was worked around by subsequent processing of the transmitted
timestamp.

Fixes: #44997
Releases: master, 2.0
Change-Id: Ia26d68945dd6deae9438de7ba71dee67cff4626d