[BUGFIX] Remove unneeded filesystem calls
This change removes a call to is_dir in the FileMonitor
which is duplicate as the ReadDirectoryRecursively does
exactly the same and throws an error in this case.
Also 2 calls to file_exists are removed as they're not
necessary at this point as filemtime() will just return
FALSE if something goes wrong.

Change-Id: I93c624cf5d5684d0fe33e8fdf348c535b742d54a
Resolves: #46290
Releases: master, 2.0