[FEATURE] Static compilation of compile time data
Implement a CompileStatic annotation for static methods that will be
evaluated by the proxy builder in production. Update bootstrap and some
classes to enable efficient caching of reflection information and lazy
loading of the reflection service.

Note: It is not possible to give a type hint to those static methods,
since it could be a DependencyProxy that is actually passed.

Change-Id: Ic29505a98a04091320294c5e389cc4e09d514fcd
Related: #46371
Releases: 2.0, master