[BUGFIX] Make the AbstractExceptionHandler check for error_reporting()
When the shut-up operator is used in PHP the error_reporting level is
temporarly lowered to 0. The AbstractExceptionHandler does not check
for this, and outputs the exception whenever it is called.

This is a problem when the set_exception_handler method is used, as
PHP will still call those exception handlers even if the shut-up
operator is used. This is why the AbstractExceptionHandler should
check for error_reporting() === 0 and ignore the exception if so.

This change fixes the exit code 255 error during the compile run
which is thrown in the Testing context where we set E_WARNING to
be an exceptional error.

Change-Id: I7c7ef5c238792183bad915fc55c3620a35cb1222
Resolves: #46485
Releases: master, 2.0