[!!!][BUGFIX] Translation not working with single numeric argument
This removes the option to provide the plural quantity as a single numeric
argument, as this was used nowhere but prevented translations from working
with such single arguments.

This change is breaking for cases where an application depends on previous
behavior that translateBy* calls with a single numeric argument would implicate
a quantity for plural form. Also classes extending I18n\Translator and using
protected method getPluralForm() will break due to a signature change.

Change-Id: I12e0856e7251684e6240c4350f4571c2207d769b
Fixes: #45062
Releases: master, 2.1