[BUGFIX] Support BCrypt validation of hashes with different cost
In the current implementation of the BCryptHashingStrategy a password is
hashed with crypt and the hash contains the algorithm and parameters
with the salt that was used to hash the password.

This change updates the validation to also take the cost from the stored
hash and allow changes to the cost setting.

Change-Id: I7dcc1425c06e3e542b545fad367a1d91d6a65689
Fixes: #47725
Releases: master, 2.0