[BUGFIX] Session storage has poor performance for hundreds of sessions
Fixes an issue with the storage backend for sessions which resulted in
a bad overall performance when hundreds of sessions were active
simultaneously. The improved implementation is now based on two separate
cache storages and a new garbage collector which allows for more fine
grained tuning.

If you previously used a different than the default cache backend for
sessions, you will need to adjust your configuration because this
implementation is based on two distinct cache backend configurations
(see updated Caches.yaml).

The option name for setting the garbage collection probability has
changed: the settings now come with a dedicated section for garbage
collection which contains the probability and, added by this patch, the
maximum number of old sessions which may be removed by the garbage
collector during one invocation.

You need to flush caches with --force after applying this change set.

Change-Id: I08578b39bab06ffb406a95f289eeb170542f92cf
Fixes: #49189
Resolves: #49190
Releases: master, 2.0