[TASK] Update and fix CGL, add PDF to Flow package
Updates and fixes the CGL to be on par with the master version, only
the namespace handling is omitted as it is not supported in 2.0.

Includes the "TYPO3 Flow Coding Guidelines on one page" PDF
as asset to the Flow documentation, fixing the download.

This also links the CGL thumbnail with the PDF and moves images
two levels up in the directory structure.

Change-Id: I93455b53e23912c44a2b45b56a4039d3c12e5356
Releases: 2.0