[BUGFIX] Fix validation if nested properties of same type are present
The performance optimization in #46340
changed the isValid() method of the
GenericObjectValidator to use the result
property while recursing through the properties
of the to be validated object.

Since GenericObjectValidator is reused, it could
be the case that this result set is reset to
empty value.

Revert this change and set the result property
*after* all properties of the to be validated
object are validated.

Resolves: #50585
Related: #46340
Releases: 2.0, master
Change-Id: I178ac02e7bcfc9748b9da4a74f51314a34717e66