[BUGFIX] Injected properties are not available in initializeObject()
This fixes an issue where properties injected to the parent class A
were not available when the initializeObject() method of sub class B
was called.

The root cause of this bug was that, in order to avoid double injection,
injectProperties() was only called in the constructor of sub class B
and not in the constructor of parent class A. The initializeObject()
methods however, were still executed in both constructors.

Change-Id: I7fbe4fc61029af60924356f49aef4964dbb346f7
Resolves: #43659
Releases: 2.0, master