[FEATURE] Respect implementation of JsonSerializable interface
The JsonView will call jsonSerialize() in transformValue()
for objects implementing the JsonSerializable interface. This is useful
if a domain model or data transfer object needs custom serialization
logic for JSON.

The implementation also works in PHP 5.3 where the JsonSerializable
interface is not yet present and will be ignored.

Change-Id: I77936ed284c55d0549b71b3739b76491bbdd9bb6
Releases: master, 2.3
Resolves: #54092