[!!!][BUGFIX] Sort packages by dependency
Before this, packages were sorted with PHP sort functions which did
not work in all cases due to the insufficiently determined order of
package dependencies.

With this change the sorting is done with an depth-first algorithm
that makes sure that package settings overrule settings from
depending packages.

The algorithm is an adapted version of
http://en.wikipedia.org/wiki/Topological_sorting

This is a breaking change in case you relied on the previous (and
sometimes incorrect) sorting of packages.

Change-Id: I2c948e7a0d5b204f2a40ea2c78bb9db256209e91
Releases: master, 2.1