[BUGFIX] .htaccess is overwritten by the composer install scripts
The composer update / install scripts overwrite the Web/.htaccess
file which is unwanted behavior as this leads to issues when
people add custom rewrite rules, or if for example setting the
MultiViews option is not allowed.

This change moves the .htaccess, index.php and flow scripts
to the Defaults folder so it will only be initialy placed into
the installation.

Change-Id: Id6233a9694b5ecbc85f81b8f6d95f96412802347
Fixes: #54553
Releases: master, 2.1