[BUGFIX] Fix support for absolute URIs in WebRedirect options
Fixes support for authentication provider redirects to external
domains.

Background:

If a ``WebRedirect`` is configured with a relative URI like::

  entryPoint: 'WebRedirect'
  entryPointOptions:
    uri: 'some/path'

The ``Location`` header gets properly prefixed with the current
base URI.
But due to a typo this happened as well for absolute URIs.

Note: using the "uri" option for internal redirects is deprecated,
use the ``routeParts`` option instead for those.

Change-Id: I203459da70465d78d50d226d265c88224ac50963
Fixes: #56598
Releases: master, 2.2, 2.1