[BUGFIX] AOP works with __clone call on parent objects
If, for example, you extend an entity which implements __clone
the AOP Framework breaks with an warning in development mode,
that it cannot access Flow_Aop_Proxy_targetMethodsAndGroupedAdvices.

Solution is to check  if the private
Flow_Aop_Proxy_targetMethodsAndGroupedAdvices property is
accessible and otherwise skipping the Advice call.

Change-Id: I96f7d15acb0d1149dac958b35218695ed0c744b2
Related: #37571
Releases: master, 2.3, 2.2, 2.1