[BUGFIX] Proxy compiler should only remove opening php tag
Without this change the proxy compiler throws away the full line
containing the opening php tag, but that is not correct as this
first line could contain important information like the namepace
declaration.

This change just removed the opening php tag and all following
whitespace but leaves anything else in the line unchanged.

Change-Id: I42867963468c23a3dd11512c31ab1bff0ba3db1c
Releases: master, 2.2, 2.1
Fixes: FLOW-3