[BUGFIX] Properly resolve case of Subpackage Key in ActionRequest
``ActionRequest::getControllerSubpackageKey()`` failed to return the
correctly cased subpackage key.
This is not the case for the other ``getController*()`` getters and
can lead to issues (e.g. "Template could not be loaded" Fluid
exceptions on case-sensitive file systems.

This change adjusts the ``getControllerSubpackageKey()`` method to
``getControllerName()`` which already uses the correctly cased
controllerObjectName to extract the controller name.

Change-Id: Ie4e4ab444eeff7396ee3f9d019fa7cd6a382ea4d
Fixes: FLOW-126
Releases: master, 2.3, 2.2, 2.1