[BUGFIX] Enable Functional Tests for PSR-4 packages
Packages with PSR-4 autoloading will not be able to work with
functional tests because two problems arise. First the
package namespace is prepended twice and second the path
contains an unnecessary backslash.

To avoid wrong amounts of backslashes we now use the path merging
utility to merge namespace parts instead of doing simple string
concatenation. Additionally a check is added for packages with
PSR-4 autoloading enabled.

Releaes: master, 2.3
Change-Id: Ic32b0720e93a5be49e8d76f64d43af8a704e0c60