[BUGFIX] Rename redirectToReferringRequest() to forwardToReferringRequest() in ActionController
``ActionRequest::redirectToReferringRequest()`` actually triggers a
forward not a redirect.
This change deprecates the ``redirectToReferringRequest()`` method in
favor of a new method ``redirectToReferringRequest()`` which works
exactly like the previous method.

The deprecated method might be re-implemented to issue a "real" HTTP
redirect at some point, so you should not use it for forwards.

Change-Id: Ife430f8e277fe39c3f5930f41095e7d6e5fe72d7
Releases: master, 2.3
Fixes: FLOW-164