[BUGFIX] Fix support for PHP 5.3
This is a follow up to the "[BUGFIX] Disable security for CLI requests"
(I1dc9d463bbe651f5cfc4faa646b35ae8bac85fe5) that introduced a regression
for PHP versions < 5.4.

Note:
This is only applied to the 2.3 branch because the regression hasn't
been merged in older branches yet and because 3.0 relies on PHP 5.5
anyways.

Change-Id: Ib62677fb51702a6f80a4b292c1e1e6060f0afb28
Releases: 2.3
Related: FLOW-163