[BUGFIX] Simplify and stabilize security context hash calculation
This simplifies ``Security\\Context::getContextHash()`` and makes sure
that it never returns a hash over an *empty* string which could be the
case before.

Background:
The Security Context Hash used to be generated by iterating over a
list of "ContextHashComponents". Those could be influenced using the
``setContextHashComponent()`` method.
The drawback of this approach is a pretty complex mechanism that is not
easy to comprehend. Besides resetting of the hash was error-prone and
could lead to hash components being ignored.

This also removes the method ``getRolesHash()`` that was neither
documented nor used outside of the ``Security\\Context``.

Change-Id: I2547e6276860aea3906755ac9d74df8b94ae5baf
Related: NEOS-5
Related: NEOS-433
Releases: master, 3.0