[BUGFIX] Disable authorization checks in functional tests
This adjusts the ``FunctionalTestCase`` base class to disable
authorization checks if the ``testableSecurityEnabled`` is not set.

Background:
If security is not enable for functional tests, the
``Security\\Context`` is not initialized. This leads to exceptions in
case the ``PolicyEnforcement`` kicks in. Because functional tests
inherit the configuration of the default contexts, this happens very
likely, especially when making use of node related privileges in Neos
sites.

Change-Id: Ie5977200368f65734a6b271357ef21306912cb62
Related: NEOS-5
Releases: master, 3.0